Re-enable exporting to CSV
