+++ tools/prepareLandscape.js
@@ -2,6 +2,7 @@ import path from 'path'
 import { safeLoad } from 'js-yaml'
 import { readFileSync, writeFileSync, mkdirSync, existsSync, rmdirSync } from 'fs'
 import { execSync } from 'child_process'
+import prepareItemsForExport from './prepareItemsForExport'
 
 const projectPath = process.env.PROJECT_PATH || path.resolve('../..')
 const settingsPath = path.resolve(projectPath, 'settings.yml')
@@ -21,3 +22,6 @@ if (!existsSync('./public/data')) {
 items.forEach(item => {
   writeFileSync(`./public/data/${item.id}.json`, JSON.stringify(item))
 })
+
+const itemsForExport = prepareItemsForExport(items)
+writeFileSync(`./public/data/items-export.json`, JSON.stringify(itemsForExport))
